Start a refactoring into smaller modules.
- Expose a Bookie api vs the raw post/get handlers. This makes it easy to mock
out the bookie api bit with solid return values/callback callers. It also
reads nicer in the non-api modules.
- Pull the code into smaller modules. Try to split the bindings of things out
of main so it's leaner and easier to read what the parts are.
- Lint

Note: Incomplete and not able to run/test. There needs to be an event system
between the modules to help them talk when prefs are changed or calls to
bookmark are successful.